Complete CLI architecture implementation per specification
Restructured CLI according to docs/dev/cli-architecture.txt:
- Moved main.go to cmd/main.go (correct location per spec)
- Moved CLI files from treex/internal/cmd/ to treex/cmd/
- Simplified api.go to thin dispatch layer as intended
- Moved core tree building logic to treex/tree.go
- Implemented default command behavior (naked "treex" works same as "treex tree")
- Fixed formatNumber function in rendering package
- All 256 tests passing with 71.7% coverage

Architecture now correctly follows separation of concerns:
- CLI layer: argument parsing, help, error handling, output rendering
- Core API: business logic via thin dispatch in api.go
- Tree building: main logic in tree.go with structured input/output
- Rendering: format-specific output generation
